auth_service/requirements.txt
Flask==2.3.3
PyJWT==2.8.0
pytest==7.4.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
redis==5.0.1
fakeredis==2.20.1
Flask-Migrate==4.1.0